[BUGFIX] Except when 'renderType' is not string in NodeFactory
In cases where `array_merge_recursive` is used to append
to a TCA configuration multiple times with the same
key, it will lead to transforming an array key with
a single value to an array of values instead.

The NodeFactory can not deal with this case.

This patch throws an exception in case no string
is found, and tries to hint at this possible
array scenario.

Resolves: #106177
Releases: main, 13.4, 12.4
Change-Id: I98ce2bafeabed23cfff6246f4f56918991366396
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88288
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>