[TASK] Simplify `FormEngine.getFieldElement()`
The method `FormEngine.getFieldElement()` is simplified with this
commit. The form element is not queried anymore everytime from DOM.
Instead, `FormEngine.formElement` is used which in turn uses internal
DOM references. The same is done for retrieving the actual field via
`namedItem()`.

The appendix `_hr` is not used anymore since TYPO3 v7 and is therefore
removed.

Resolves: #106266
Releases: main, 13.4
Change-Id: I65d8032eaea2f258529df597b264f3d4a7a8b38c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88384
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Andreas Nedbal <andy@pixelde.su>