[TASK] FormEngine: Avoid unneeded regexes when parsing exclusive values
The method `FormEngine.setSelectOptionFromExternalSource()` takes care
of so-called "exclusive values" – values, that are mutually exclusive in
usage, as e.g. seen in `pages.fe_group`.

Exclusive values are passed as a comma-separated value. When an incoming
value is checked for existence in this CSV, multiple regular expressions
are composed, which is quite an overhead.

Exclusive values are now passed as an array, where value existence can
easily be checked via `.includes()`.

The method signature of `setSelectOptionFromExternalSource` is changed
in this patch. The code search didn't reveal usage in any TER extension,
therefore this is not considered being a breaking change.

Resolves: #106277
Releases: main, 13.4
Change-Id: I611c01b977f16df6cee6e824760143b658411a3c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88420
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>