[BUGFIX] Streamline language switch in columns only mode
In "columns only" mode, only a subset of available fields
are displayed for a record. Those views are usually used
for bulk editing or editing just specific context related
fields, e.g. the pages seo data. Those views are not intended
for localization as this does not work with "columns only".

Therefore, the language switch does no longer propose to
translate the record to another language, when in "columns
only" mode.

But: The language switch does now respect the "columns
only" config for existing records. This means switching
to a translation keeps the configured fields.

Resolves: #103853
Releases: main, 13.4
Change-Id: I3707c590029e2940fa5398996f72bab79d44a555
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88434
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>