[BUGFIX] Remove duplicated event listener
Multi record selection comes with default
event listener JS modules for the "edit"
and "delete" actions.

The Recordlist defines its own event listener
to additionally handle the "columns only"
mode, which allows editing multiple records
with just a subset of fields.

Therefore, the default JS module is now
removed as it interfered with the custom
logic, leading to the strange behavior where,
depending on the loading order, the "columns only"
mode sometimes worked and sometimes not.

Resolves: #106281
Releases: main, 13.4, 12.4
Change-Id: Iaca58a2d9e72e851a17714d2e6ea516513328675
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88411
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Oliver Bartsch <bo@cedev.de>