[BUGFIX] Avoid layout shifts when dragging content blocks in Page module
We now use appropriate move icons to indicate the desired actions and
visualize this in the tooltip. Previously, information about pressing
"Control" to copy an element instead of moving it caused layout shifts.
This issue arose because the information was not migrated correctly and
was previously attached to the drag preview.

With the adoption of the new DragTooltip API, we ensure a more natural
behavior. Instead of displaying the help text, we update the cursor and
status icon in the tooltip to reflect the desired action. This approach
aligns with the standard behavior on operating systems.

Additionally, the detection of whether a copy action is requested has
been adjusted to work with the Option key on macOS. This is now also
properly forwarded to the modal to actually perform the requested action.

Related: #100258
Resolves: #106319
Releases: main, 13.4
Change-Id: I58e25b839188af0dcb603cb6c266ee1423dc7b28
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88549
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>