[BUGFIX] Dispatch `multiRecordSelection:checkbox:state:changed` once
The module `@typo3/backend/multi-record-selection` uses multiple events
to notify about state changes. Unfortunately, it happened that the event
`multiRecordSelection:checkbox:state:changed` was dispatched twice when
the checkboxes are toggled via the action dropdown:

* explicitly in `changeCheckboxState()`
* by the `change` event listener registered in
  `registerDispatchCheckboxStateChangedEvent()`

The explicit call in `changeCheckboxState()` is removed, as the `change`
event is triggered anyway and it's event handler is also called when
single rows get (un)checked.

Resolves: #104833
Releases: main, 13.4, 12.4
Change-Id: I03622acc96b7466189c5f2315265ea5ee5b3c797
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88618
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>