[BUGFIX] Register global keydown handler to close modals
Usually, closing modal windows via pressing <ESC> works just fine as our
modals are based on Bootstrap. However, this may fail if a focussed
element within the modal is removed from the DOM, as the browser may
change the focus to the outer `body` element then.

To circumvent this, the Modal API will install a global `keydown`
handler that checks whether <ESC> was pressed and closes the modal then.
The event handler itself gets removed once the modal is finally closed.

Resolves: #102329
Releases: main, 13.4
Change-Id: Ie82236db5ee85ee592cf609890fc489816f2c75a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88965
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>