[TASK] Abort field initialization early in FormEngine validation
The FormEngine Validation module needs to initialize its input fields
before being able to validate them. This actions needs to be done in
case new fields are added, e.g. in IRRE. To avoid re-initializing
fields, a data attribute is set.

The code now checks early whether the data attribute is set to avoid
parsing JSON and querying the DOM which may not be needed at all.

Resolves: #106531
Releases: main, 13.4
Change-Id: Ibac5d5d99507e744f46e6cadcc8bbbcbecf602ba
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89028
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>