[BUGFIX] Properly link to moved CEs in workspace
When somebody links to a content element in a
workspace, one should always link to the LIVE version
of the CE. This is adapted.

In addition, CEs which have been moved in a workspace,
are now only shown ONCE and in the right position
by using BackendUtility::workspaceOL() and its fourth
argument properly.

Resolves: #104504
Releases: main, 13.4
Change-Id: Ibbe04532842a044cb86f24f808ae1dcb0285f9e0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89009
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>