[BUGFIX] Check state in web components extending `HTMLElement`
Web components that extend `HTMLElement` directly suffer from poor
lifecycles as `connectedCallback()` is called when the component is
added into the DOM, and when it is moved. This causes several issues,
e.g. double initialization or stacked event listeners.

This is a major issue in FormEngine as IRRE elements may get re-ordered
and thus the position of their children change in the DOM, calling
`connectedCallback()` again.

To solve this issue in FormEngine context for now, internal state is
checked and futher processing is aborted, if suitable.

Resolves: #103815
Releases: main, 13.4, 12.4
Change-Id: I74bfcda1538492bc6323d6a3782408741b9ca50c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89201
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>