[BUGFIX] Respect placeholder configuration in RTE fields
This change streamlines the existing TCA placeholder handling
of type=text to also work for RTE fields (that means when
`enableRichtext` is set), to align with existing input and text
field handling.

It is configured like regular input/text field TCA
(see EXT:styleguide RTE field 'rte_6'):

```
// ...
  'rteColumn' => [
    'label' => 'Some RTE field',
    'config' => [
      'type' => 'text',
      'enableRichtext' => true,
      'default' => null,
      'nullable' => true,
      'mode' => 'useOrOverridePlaceholder',
      'placeholder' => 'Insert contact details here.',
    ],
  ],
// ...
```

Inherited HTML for the placeholder (IRRE) is evaluated as proper
markup within the RTE. Labels only show the "strip_tags()"
variant of possible HTML context.

The logic for displaying placeholders and "defaults" is a bit
unfortunate, because what the UI shows as "Default" is not really
a default, but instead an inherited fallback for a decoupled
rendering process that acts on NULL values.

We should have three terms:

* Placeholder = Something the editor reads and knows what
  to fill in ("Enter file description")
* Default = A value that is actually saved ("NULL" to trigger
  fallback for example, and otherwise the actual string content)
* Fallback = A value that will be used in case the stored value
  is set to NULL

This is something that is to be streamlined across all TextElements.

The patch is meant to make RTE functionality available so that
for example `sys_file_reference` fields can be remapped to
utilize richttext editing and still keep their placeholder/fallback
functionality in place.

Resolves: #106636
Releases: main, 13.4
Change-Id: Ie0f519f75561f30e0663c1a0f00eb0cd694e0eb7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89948
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>