[TASK] Improve Progress Tracker Component
Based on feedback, the progress trackers,
especially in modals, feel like a blocking
element, and users are unable to find the
navigation buttons easily. To resolve this,
we are moving the element to the top to inform
users earlier about the current progress and
to make the stepping options more recognizable.

The current implementation also limits the number
of possible steps, as each must display its label.
We are reducing information overload by showing
only the current step’s name. Additionally, we now
provide a simple numeric overview indicating the
current step and the total number of steps.

Resolves: #106706
Releases: main, 13.4
Change-Id: If93c3c8b4ff5529b0d534c89620d445d077af179
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89512
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>