[BUGFIX] Allow language switch in "Open in new window" popup
Prevent JavaScript errors when switch languages
in FormEngine popup mode ("Open in new window").

The `<typo3-backend-module-router>` element is not available
in this mode, which is why the desired URL is assigned
to the frame directly.

Also sudo mode is adapted to avoid the workaround that has
been added in #106771, since @typo3/backend/viewport/content-container
is now handling this case.

This patch ensures the wanted location switch can be performed in
this case. An e2e test (playwright) is added to verify this.

Resolves: #106718
Resolves: #104302
Resolves: #102270
Related: #106771
Releases: main, 13.4
Change-Id: If942816660b953fe06fdf707ad27ac18fd94613e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89429
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Garvin Hicking <garvin@hick.ing>