[TASK] Remove entrypoint "typo3/install.php"
This change removes the separate entrypoint
"typo3/install.php" from TYPO3 Core.

Instead, the Install Tool is now routed to
the TYPO3 Backend system instead of the file
"/typo3/install.php", so it follows the
$GLOBALS[TYPO3_CONF_VARS][BE][entryPoint] option,
and the maintenance mode can then be available
in other locations - appending /install.

If the system is not set up properly - or
no .htaccess / rewrite rules - the GET variable
"__typo3_install" is used to access the install
tool directly.

Resolves: #107536
Releases: main
Change-Id: I3e4832155a6a9984ff2289e0e9d2eab76a966142
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89503
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>