[TASK] Move settings-editor labels to dedicated file
Currently the labels for the site settings module
and the ones for the settings editor are mixed and
need to be splitted. The settings-editor can be
used independant outside site settings module, to
enable this intended behavior we are moving the
language labels for that editor to a dedicated
file.

Resolves: #106781
Releases: main, 13.4
Change-Id: I47c4d86da89628958c5230d7fd08885b436a46aa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89565
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>