[BUGFIX] Fix negated disallowed check in drag-uploader
`checkDisallowedExtensions()` is expected to return
false if the extension is disallowed, the negation was
incorrectly migrated with #102415.

This mistake is hard to guess, given the ambiguous name,
therefore the methods for allow-list and disallow-list
are now renamed to reflect their intended result and
the conditions are adapted to check the particular
cases.

Resolves: #106784
Related: #102415
Releases: main, 13.4
Change-Id: Ie9453e4e9f23efb89b550ae462d26986ffe15dca
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89588
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>