[BUGFIX] Add installation method to system information toolbar
Replace the boolish "Composer mode" entry which only shows "Enabled" if
the system is managed via composer with a generic "Installation Method"
entry.

The previous entry failed to inform "Classic mode" users about their
installation environment and prevented users to quickly detect if
a TYPO3 instance is running in Classic mode, as it is hard to find
a missing entry and realize that the lack of information means
"Classic mode".

Releases: main, 13.4, 12.4
Resolves: #106932
Change-Id: Ic34e916691f012076587c2acab271de4ffa101b7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89805
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>