[BUGFIX] Prevent modals from closing on Escape in search
When pressing Escape inside an input[type="search"],
browsers clear the input value by default. However,
Bootstrap modals also listen for Escape globally and
close the modal, which is unintended in this case.

This change stops Escape key propagation when a
search input is focused, allowing only the input to
clear, but keeping the modal open.

The event listener is registered in the capture phase
to ensure it runs before Bootstrap's internal handler.

Resolves: #106943
Releases: main, 13.4
Change-Id: Ic9c0a8dcc3b1e9e16cd2f692a00b071ef76a5884
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89833
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Oli Bartsch <bo@cedev.de>