[BUGFIX] Allow recursive restore of records via recylcer for editors
The current approach to lookup deleted records in the recycler module
is to apply a fragile runtime TCA modification that effectively removes
the `delete` ctrl from pages in order to make rootline based
`isInWebMount` checks work.

This works in most cases, but there are corner cases where recursive
records are checked, and `BackendUserAuthentication::isInWebMount`
uses `BackendUtility::getPageForRootline` to obtain a rootline
both for deleted content elements and for pages, but it does only
apply the TCA modification when processing pages, not when checking
content elements. This results in an invalid (runtime) rootline cache
being generated/used when creating the list of deleted pages *and*
content elements.

Instead of faking TCA (which will require schema rebuilds once we
migrate to TCA schema) or extending the rootline cache to hash current
schema/TCA  the relevant access checking methods now learn to ignore the
delete-restriction instead of invalidly and globally modifying TCA.

Releases: main, 13.4
Resolves: #93714
Related: #70863
Change-Id: I5dfd9deae1573cc6ee8765fb8ba23fc54898a05e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89969
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>