[TASK] Replace $_FILES with PSR-7 UploadedFile in ExtendedFileUtility
This change substitutes the low-level usage of super global
`$_FILES` with the corresponding PSR-7 `UploadedFile` models.

The `is_uploaded_file()` check in `ResourceStorage` is only
performed for path values (as string). In case `UploadedFile`
structures are given, the assumption is, that file were actually
uploaded (this allows functional testing via `UploadedFile`).

Resolves: #107027
Releases: main
Change-Id: I11f5cba8542f92a73273320f074eff13d8b20a27
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89983
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>