[TASK] Allow using CKEditor in fullscreen mode
The CKEditor used in the TYPO3 backend now utilizes its "Fullscreen"
plugin [1], allowing to open CKEditor in a fullscreen view. To allow
editors to reach important actions nevertheless, the module's document
header will stay visible.

The CKEditor module `@ckeditor/ckeditor5-fullscreen` is installed to
enable toggling the fullscreen view in CKEditor.

The module is enabled by default in the `Full.yaml` preset, acting as a
"show off" configuration that should not be used in production.

To enable the module, the toolbar item `fullscreen` needs to be added.

Executed command:

    npm install --save @ckeditor/ckeditor5-fullscreen@^45.2.1
    npm run build

[1] https://ckeditor.com/docs/ckeditor5/latest/features/fullscreen.html

Resolves: #106962
Releases: main, 13.4
Change-Id: I8da40b733ca131d8e0a3b3135c9f6f4db5dbcef0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90062
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>