[BUGFIX] Ensure setting input fields do not lose focus on value change
Make sure that the settings specific type field is always
rendered and never temporarily replaced by a loading spinner,
which causes input focus to be lost when changing values via
keyboard (e.g. by pressing up or down in a `number` field).

Note: The respective rendering is triggered because the has-change
class was to be added, but caused this glitch as a side effect.

Resolves: #107091
Releases: main, 13.4
Change-Id: I82395cdff041503ef344aa5161d32f848f78cbf4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90074
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>