[BUGFIX] Do not cast PID of file reference to int in HTML
The value of the PID of a new file reference may not be
an integer, but a string with a "NEW..." reference. This
can happen if a file is added to a page field, like "media",
when that page was not saved yet.

Instead write the value as as string, with htmlspecialchars
applied. This is the same behaviour as the InlineRecordContainer

Resolves: #103092
Releases: main, 13.4, 12.4
Change-Id: I13c56b036ee84d25a83693a2d6a386e0777d44d1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90474
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>